abstracting out common code that is compatible for both PLAY 2.6 and 2.7
